+++ examples/sending-plain-text-email-example/src/main/java/com/test/PlainTextEmailApplication.java
@@ -1,9 +1,13 @@
 package com.test;
 
+import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.support.AbstractApplicationContext;
 
 import javax.annotation.PostConstruct;
 import java.io.UnsupportedEncodingException;
@@ -13,13 +17,20 @@ import java.util.concurrent.TimeUnit;
 
 @SpringBootApplication
 @ComponentScan(basePackages = {"com.test", "it.ozimov.springboot.templating.mail"})
-public class PlainTextEmailApplication {
+public class PlainTextEmailApplication implements ApplicationContextAware {
+
+    private ApplicationContext applicationContext;
 
     @Autowired
     private TestService testService;
 
     public static void main(String[] args) {
         SpringApplication.run(PlainTextEmailApplication.class, args);
+    }
+
+    @Override
+    public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
+        this.applicationContext = applicationContext;
     }
 
     @PostConstruct
@@ -33,7 +44,7 @@ public class PlainTextEmailApplication {
         TimerTask shutdownTask = new TimerTask() {
             @Override
             public void run() {
-                System.exit(0);
+                ((AbstractApplicationContext) applicationContext).close();
             }
         };
         Timer shutdownTimer = new Timer();