+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/templating/mail/service/ServiceStatus.java
@@ -2,7 +2,8 @@ package it.ozimov.springboot.templating.mail.service;
 
 public enum ServiceStatus {
 
-    CREATED,
+    RUNNING,
+    CLOSING,
     CLOSED
 
 }